pyproject.toml
[tool.poetry]
name = "mainapp"
version = "0.1.0"
description = ""
authors = ["xshadowbringer <sh4dowbringer10@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pymycobot = "^3.4.4"
pyusb = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
